+++ lib/debug.dart
@@ -1,2 +1,1 @@
-export 'package:token_parser/token_parser.dart';
 export 'package:token_parser/src/debug.dart';